Use the global $table_prefix in switch_to_blog()
When using batcache in combination with memcached object cache on a single site install, table_prefix is not available when switch_to_blog is called. This causes two versions of data to be stored in memcached.

By bringing in the global, we can be sure that table_prefix is consistent in both logged in and logged out states